chore(src): rename util -> utils
